dev_ns__adopt_the_android_alarm-dev_driver
Add device namespace logic to Android alarm-dev driver - so processes in a
namespace would only see alarms in its own namespace.

Place per-namespace data into struct alarm_dev_ns, and convert the driver
to operate on the current process's context.

The exception is an ioctl call with ANDROID_ALARM_SET_RTC, in which case
the fact that the RTC setting changes must be communicated to all of the
contexts (device namespaces) so it takes effect in each namespace. This is
by calling the new propagate_alarm_time_change().

Changelog:
[16-Aug-2013] v1 - initial version

Change-Id: I968989e23268d9510f40cf90b7c3366c03d4a226
Signed-off-by: Oren Laadan <orenl@cellrox.com>
Signed-off-by: Amir Goldstein <amir@cellrox.com>