Replace reference of pid_ns with init pid in dev_ns
This closes Cellrox/devns-patches#4

The dev_ns is coupled to pid_ns at creation time. There is no need for
dev_ns to keep a reference to pid_ns. The reference was used to be able
to tell the pid of the respective init task for use by printk; instead,
store the pid itself in dev_ns.

Change-Id: I4997582ef7cf9cdde5b9919d80d06ea058b6b561
Signed-off-by: Oren Laadan <orenl@cellrox.com>
Reported-by: Jeremy Andrus <jchristian.andrus@gmail.com>