dev_ns: introduce device namespaces
Introduce device namespaces:

The idea with a device namespace comes from the Android-Cells project where
namespaces are utilized to create a container-like environment on Linux, only
where there's a notion of an 'active' namespace and all other namespaces are
inactive. In such a case only processes residing within the active device
namespace should communicate with actual devices, where processes inside
inactive containers should be able to communicate gracefully with the device
driver, but not the device.

Changelog:
[11-Sep-2014] v2 - fix compile and crashes with !CONFIG_DEV_NS
  - replace reference of pid_ns with pid_t in dev_ns [report: Jeremy Andrus]
  - fix compile errors when !COFNIG_DEV_NS [report: Jeremy Andrus]
  - fix kernel crash when !CONFIG_DEV_NS [report: Jeremy Andrus]
[16-Aug-2013] v1 - initial version

Change-Id: I38ef9fdf76086794d42679f4ec05612ba0cfb179
Signed-off-by: Oren Laadan <orenl@cellrox.com>
Signed-off-by: Amir Goldstein <amir@cellrox.com>